Update cast in msg.set (#1130)
A while ago, MySensors changed to use int32_t
instead of long int (and similar for other data
types). This change was not applied to all examples,
causing some examples to fail compilation for
ESP8266 with the following message:

call of overloaded 'set(long int)' is ambiguous

This update changes the casts to match the new style.
This fixes
https://github.com/mysensors/MySensors/issues/1129
Thanks to https://github.com/henfri for reporting.

+++ examples/SoilMoistSensor/SoilMoistSensor.ino
@@ -138,7 +138,7 @@ void loop()
 	Serial.println ();
 
 	//send back the values
-	send(msg.set((long int)ceil(sensor1)));
+	send(msg.set((int32_t)ceil(sensor1)));
 	// delay until next measurement (msec)
 	sleep(SLEEP_TIME);
 }
